Added ObjModelBoundaryFinder2 as temporary fix for buggy ObjModelBoundaryFinder. Removed aliasing of smart pointer in DijkstraShortestPathFinder. Removed buggy ObjModel::clone() method. Use static ObjModel::copy()instead. Raised upper limit on texture resolution in ObjModel material addition.

+++ include/template/VectorFloatKeyHashing_template.h
@@ -29,7 +29,7 @@ Key<T,M>::Key( double v[M], int pw)
 template <typename T, int M>
 Key<T,M>::Key( float v[M], int pw)
 {
-    const float m = powf(10,pw);
+    const float m = powf(10.0f,(float)pw);
     for ( int i = 0; i < M; ++i)
     {
         _ielems[i] = (T)roundf(v[i]*m);
